Create code 5.0 (Under Construction)
fixed the adding issue that kept messing up the score.  the gamin continues with out any adding issue. 

Still: 
need to fix random error that populates in the "DEAL" are for some reason at certain times it states the that index is out of list meaning something is going initially or through out code.