Create Construction 6.0
fixed the random "out of list" error that appears at random times.  The issue:  I confused the syntax of a random.randint with the for loop range.  meaning the for loop range discount the inputed parameter i.e.  (0, 10) = 1 - 9

Where in random.randint(0, 10) equals  = 0-10 which i need to denote a -1 to help the code reference the list [0 -9] instead.  every time random.randint chosed 10 it was out the range of the list causing the app to crash. 

All I need to do is fix the 1 vs 11 in the As card in the list. 
most likely if 11 exist and > 21 then (-10) the calculation so that there is a chance the code can continue.