Create Final product ver 3.0
This is the final version as a launching product. 
fixed the issue where the computer was not accounting for an As card.  meaning it would think all A's were equal to 11 when A's can also be equal to 1.  

At the moments there isn't any bugs to account for.  This would be a beta.